\name{SwathToVec}
\alias{SwathToVec}
\title{RSWATH (Smart Word Analysis for THai)}
\usage{
  SwathToVec(inputfile)
}
\description{
SwathToVec reads the output file from rswath and return as a vector.
}
\examples{
SwathToVec(inputfile = "mytextfile.txt")
}
\author{
Sompob Saralamba (saralamba@gmail.com)
}
